requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
python-jose[cryptography]
requests
python-dotenv
